requirements.txt
torch==2.8.0
torchvision==0.23.0
fastapi==0.116.2
streamlit==1.46.0
pillow==11.2.1
uvicorn==0.35.0
